# perf(bin/server): increase msg size and don't allocate msg per resp